Changed some uses of some integer and float types.

+++ src/main.cpp
@@ -8,12 +8,12 @@
 
 int main() {
     using namespace std::chrono_literals;
-    int64_t start_val = 0;
-    int64_t stop_val = 0;
+    uint64_t start_val = 0;
+    uint64_t stop_val = 0;
     const int repeat_val = 10;
-    double iteration_length_arr[repeat_val];
-    double avg_search_time = 0.0;
-    int64_t prime_count = 0;
+    uint64_t iteration_length_arr[repeat_val];
+    long double avg_search_time = 0.0;
+    uint64_t prime_count = 0;
     std::string choice;
     std::string yes_choices[] = {"Y", "y", "Yes", "yes"};
     std::string no_choices[] = {"N", "n", "No", "no"};
@@ -24,7 +24,7 @@ int main() {
         while (true) {
             std::cout << "Enter an integer start value greater than 0 and less than 9223372036854775807 (aka 2^(64 - 1) - 1): ";
             std::cin >> start_val;
-            if (start_val > 0 && start_val < std::numeric_limits<int64_t>::max()) {
+            if (start_val > 0 && start_val < std::numeric_limits<uint64_t>::max()) {
                 break;
             } else {
                 std::cerr << "Value out of range!" << '\n';
@@ -34,7 +34,7 @@ int main() {
         while (true) {
             std::cout << "Enter an integer stop value greater than the start value and less than 9223372036854775807 (aka 2^(64 - 1) - 1): ";
             std::cin >> stop_val;
-            if (stop_val > start_val && stop_val < std::numeric_limits<int64_t>::max()) {
+            if (stop_val > start_val && stop_val < std::numeric_limits<uint64_t>::max()) {
                 break;
             } else {
                 std::cerr << "Value out of range!" << '\n';
@@ -47,13 +47,13 @@ int main() {
         for (int k = 0; k < repeat_val; k++) {
             std::cout << "Iteration " << (k + 1) << " of " << repeat_val << " Runtime (ns): ";
             auto iteration_start_time = std::chrono::steady_clock::now();
-            for (int64_t i = start_val; i < stop_val + 1; i++) {
+            for (uint64_t i = start_val; i < stop_val + 1; i++) {
                 if (func::prime::is_prime(i)) {
                     prime_count++;
                 }
             }
             auto iteration_stop_time = std::chrono::steady_clock::now();
-            std::chrono::duration<int64_t, std::nano> elapsed_single_iteration = iteration_stop_time - iteration_start_time; //how many nanoseconds have elapsed.
+            std::chrono::duration<uint64_t, std::nano> elapsed_single_iteration = iteration_stop_time - iteration_start_time; //how many nanoseconds have elapsed.
             iteration_length_arr[k] = elapsed_single_iteration.count(); //store elapsed_single_iteration nanoseconds in iteration_length_arr.
             std::cout << elapsed_single_iteration.count() << '\n';
         }
@@ -61,8 +61,8 @@ int main() {
         avg_search_time = std::accumulate(std::begin(iteration_length_arr), std::end(iteration_length_arr), avg_search_time) / repeat_val;
 
         auto prog_stop_time = std::chrono::steady_clock::now();
-        std::chrono::duration<int64_t, std::nano> elapsed_prog_runtime = prog_stop_time - prog_start_time; //how many nanoseconds have elapsed.
-        int64_t prog_runtime_nanoseconds = elapsed_prog_runtime.count();
+        std::chrono::duration<uint64_t, std::nano> elapsed_prog_runtime = prog_stop_time - prog_start_time; //how many nanoseconds have elapsed.
+        uint64_t prog_runtime_nanoseconds = elapsed_prog_runtime.count();
 
         std::cout << "Prime benchmark is done!" << '\n';
 
@@ -72,9 +72,9 @@ int main() {
         std::cout << "Program ran for: " << prog_runtime_nanoseconds << " ns" << '\n';
         std::cout << "\n";
         std::cout << "Average time to find all primes between " << start_val << " and " << stop_val << " was (DD:HH:MM:SS.SSSSSSSSS):" << '\n';
-        std::cout << func::to_days_hours_minutes_seconds((int64_t) avg_search_time) << '\n';
+        std::cout << func::to_days_hours_minutes_seconds((uint64_t) avg_search_time) << '\n';
         std::cout << "\n";
-        std::cout << "Average search time: " << (int64_t) avg_search_time << " ns" << '\n';
+        std::cout << "Average search time: " << (uint64_t) avg_search_time << " ns" << '\n';
         std::cout << "Number of primes found is: " << prime_count / repeat_val << '\n';
     } else if (std::find(std::begin(no_choices), std::end(no_choices), choice) != std::end(no_choices)) {
         std::cout << "Exited program." << '\n';
